fix: 찜하기 UI 분리
